+++ src/globals.py
@@ -1,2 +1,2 @@
-build=41
+build=42
 type='dev'